% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\docType{data}
\name{pop}
\alias{pop}
\title{FIA Population Level Auxiliary Data for Oregon}
\format{
An object of class \code{data.frame} with 10060 rows and 10 columns.
}
\usage{
pop
}
\description{
FIA Population Level Auxiliary Data for Oregon
}
\keyword{datasets}
