% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.R
\docType{data}
\name{samp}
\alias{samp}
\title{FIA sample data for Oregon}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1494 rows and 11 columns.
}
\usage{
samp
}
\description{
FIA sample data for Oregon
}
\keyword{datasets}
